gemini-impl/backend/requirements.txt
fastapi
uvicorn
python-multipart
gitpython
pydantic